refactor: FoodProvider 추상화
